more progress

+++ commands/delete.sh
@@ -3,7 +3,7 @@
 if [ "$MCCLI_DOCKER" = "true" ]; then
 	docker ps --no-trunc | grep "$1" >/dev/null	
 else
-	[ ! -e "$data_dir/.running" ]
+	[ -e "$data_dir/.running" ]
 fi
 server_running="$?"
 
@@ -20,5 +20,11 @@ if [ "$server_running" -eq 0 ]; then
 		sleep 5;
 	done
 fi
-docker rm "$1"
+
+if [ "$MCCLI_DOCKER" = "true" ]; then
+	docker rm "$1"
+fi
+
+rm -rf "$2"
+
 echo "Deleted server $1" >&2
